a lot of changes

+++ index.js
@@ -1,12 +1,27 @@
 const { Client, GatewayIntentBits, Collection, Permissions, EmbedBuilder } = require('discord.js')
 const fs = require('node:fs')
 const config = require('./config.json')
+const mongoose = require('mongoose');
 
+const Schema = mongoose.Schema
+
+const ObjectId = Schema.ObjectId;
+
+const repSchema = new Schema({
+    userID : {
+        type : mongoose.SchemaType.String,
+        required : true
+    }
+})
 const wrong = ["me", "i", "mine"];
 const right = ["us", "we", "our"];
 
 const bot = new Client({ intents: [GatewayIntentBits.Guilds, GatewayIntentBits.GuildMessages, GatewayIntentBits.MessageContent, GatewayIntentBits.GuildMembers, GatewayIntentBits.GuildPresences] })
 bot.commands = new Collection()
+bot.cooldown = new Collection()
+bot.config = {
+    cooldown: 15000
+}
 
 for(let i = 0; i < config.folders.length; i++) {
     console.log(config.folders[i])
@@ -23,8 +38,11 @@ bot.once('ready', () => {
 })
 
 bot.on('messageCreate', async message => {
-    if(message.author.bot) return;
-    if(message.channel.type === 'dm') return;
+    let person = message.author.id
+    
+    if(message.author.bot || message.channel.isDMBased()) return;
+    
+
     for(let i = 0; i < config.banned.length; i++) {
         if (message.content.toLowerCase().includes(`${config.banned[i]}`)) {
             message.delete()
@@ -35,6 +53,7 @@ bot.on('messageCreate', async message => {
         }
         
     }
+
     
 })
 
@@ -46,7 +65,7 @@ bot.on('interactionCreate', async interaction => {
     if (!command) return;
 
     try {
-        await command.execute(interaction)
+        await command.execute(interaction, mongoEconomy)
     } catch (error) {
         console.error(error)
         await interaction.reply({ content: 'There was an error while executing this command!', ephemeral: true });
@@ -76,11 +95,16 @@ bot.on('presenceUpdate', async (oldPresence, newPresence) => {
             let newOBJ = Object.values(newPresence.activities)
             let oldName = oldOBJ[0]
             let newName = newOBJ[0]
+            
+            if(newName == 'Custom Status') newName = newOBJ[1]
+            if(oldName == 'Custom Status') oldName = oldOBJ[1]
 
-            if(!newName == config.game && !oldName == config.game) return
+            if(!newName == config.game && !oldName == config.game || (newName == undefined && oldName == undefined)) return
             if(newPresence.user.bot) return
-                    
-            c.send(`User: ${newPresence.user.username}\nNew Presence: ${newName}\nOld Presence: ${oldName}`)
+            
+            if(newName == config.game) c.send(`User ${newPresence.user.username} started playing ${newName}`)
+            if(oldName == config.game) c.send(`User ${oldPresence.user.username} started playing ${oldName}`)
+
 
                 
             
@@ -90,5 +114,7 @@ bot.on('presenceUpdate', async (oldPresence, newPresence) => {
         };
     });
 });
+
+
 
 bot.login(process.env.TTP)
